item_service/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
supabase==2.10.0
pydantic>=2.5.0
prometheus-client==0.19.0
pytest>=7.4.3
pytest-cov==4.1.0
httpx>=0.25.0
PyJWT==2.10.1
cryptography==42.0.4
requests==2.31.0
